:recycle: [StatistiqueCentre] Migrate
