--- config/doctrine/StatistiqueCentre.orm.yml
@@ -1,24 +1,0 @@
-App\Entity\StatistiqueCentre:
-    type: entity
-    table: "statistiquecentre"
-    repositoryClass: App\Repository\StatistiqueCentreRepository
-    id:
-        id:
-            type: integer
-            id: true
-            generator:
-                strategy: AUTO
-    fields:
-        nom:
-            type: string
-            length: 255
-        donnees:
-            type: array
-
-    manyToOne:
-        centre:
-            targetEntity: App\Entity\Attributes\Centre
-            inversedBy: statistiquesCentre
-            joinColumn:
-                nullable: false
-    lifecycleCallbacks: {  }
